fix issue 46 > extraced useModal
